add apply_filters function
I ran into the issue that an acf field group had multiple display rules,
like:

display filed group when it is a type of page AND
if it is a sub page of some page id

with the standard $args array of 'post_type' => 'page' the acf location
rule match would always fail as the rule check of parent_page would
always return false. and it would not display the acf field in the page
table column

with the apply_filters() i could hook into this and return an $ args
like:

array(
'post_type' => 'page',
'post_parent' => 110,  (the page id set in the acf field group display
ruke)
'post_id' => 115 ( a page id which is a child of the parent page id
)

now the acf location rule match will return true for the field group and
the acf field will get displayed in the page table column.

also ... thanks a lot for that nice piece of code. exactly what i was
looking for ;)